fixing doube warnings.

+++ newton-4.00/applications/ndSandbox/ndShadowMapRenderPass.cpp
@@ -332,7 +332,7 @@ void ndShadowMapRenderPass::RenderScene(ndFloat32)
 
 		cameraTestPoint.m_x = m_farFrustumPlanes[i];
 		const ndVector cameraPoint(cameraProjection.TransformVector1x4(cameraTestPoint));
-		m_cameraSpaceSplits[i] = ndFloat32 (0.5f) * cameraPoint.m_z / cameraPoint.m_w + ndFloat32(0.5f);
+		m_cameraSpaceSplits[i] = GLfloat(ndFloat32 (0.5f) * cameraPoint.m_z / cameraPoint.m_w + ndFloat32(0.5f));
 
 		tileMatrix[3][0] = viewPortTile.m_x;
 		tileMatrix[3][1] = viewPortTile.m_y;
